skills/pptx-skill/requirements.txt
python-pptx>=0.6.21
Pillow>=9.0.0
matplotlib>=3.5.0
requests>=2.28.0
